+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - AudienceMemberFactory::fromUserInLogic to get an audience member of a user filtered to a logic group
 - CachedAudienceMemberFactory to cache audience member results
+- Use AudienceMemberFactory::fromUserInLogic in LogicAudience
+- ActivityInstanceRepository::allForResource
 
 ## [4.4.1] - (27/04/2020)